+++ kotlin/settings.gradle
@@ -76,7 +76,7 @@ include ':add-features-feature-service',
         ':surface-placement',
         ':token-authentication',
         ':trace-utility-network',
-        ':use-geotriggers-to-show-spatially-relevant-information',
+        ':setup-location-driven-geotriggers',
         ':view-point-cloud-data-offline',
         ':viewshed-geoprocessing',
         ':viewshed-location',
