+++ kotlin/setup-location-driven-geotriggers/build.gradle
@@ -6,7 +6,7 @@ android {
     compileSdkVersion rootProject.ext.compileSdkVersion
 
     defaultConfig {
-        applicationId "com.esri.arcgisruntime.sample.usegeotriggerstoshowspatiallyrelevantinformation"
+        applicationId "com.esri.arcgisruntime.sample.setuplocationdrivengeotriggers"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode rootProject.ext.versionCode
