+++ kotlin/setup-location-driven-geotriggers/src/main/java/com/esri/arcgisruntime/sample/setuplocationdrivengeotriggers/GardenSection.kt
@@ -1,4 +1,4 @@
-package com.esri.arcgisruntime.sample.usegeotriggerstoshowspatiallyrelevantinformation
+package com.esri.arcgisruntime.sample.setuplocationdrivengeotriggers
 
 import com.esri.arcgisruntime.data.ArcGISFeature
