+++ kotlin/setup-location-driven-geotriggers/src/main/res/drawable/ic_launcher_background.xml
