+++ kotlin/setup-location-driven-geotriggers/src/main/res/layout/activity_main.xml
