+++ kotlin/setup-location-driven-geotriggers/src/main/res/values/colors.xml
